Add AVR Support to avatar-qemu for use in HALucinator.
This commit adds AVR Support to avatar-qemu. In theory,
any AVR is supported. In practice we have only tested
this with CPU Models avr5-avr-cpu, avr51-avr-cpu and
avr6-avr-cpu.

AVR has a strange representation in QEMU, using words
and not bytes. Thus the entry point needs to be
divided by two compared to its byte-representation.

Signed-off-by: Antony Vennard <antony@vennard.ch>