Debug memory writes to RAM should trigger MMIO calbacks
When using gdb to write memory mapped IO the callbacks of these IO devices
do not get triggered because it uses `address_space_write_rom` which does a
simple memcpy instead of checking if there are other things to do for that
memory.
Therefore I propose to switch to `address_space_write` if the write goes to
a RAM address.